refactor dataset type
