+++ client/starwhale/api/_impl/dataset/__init__.py
@@ -1,50 +1,3 @@
-from starwhale.core.dataset.type import (
-    Line,
-    Link,
-    Text,
-    Audio,
-    Image,
-    Point,
-    Video,
-    Binary,
-    Polygon,
-    LinkAuth,
-    LinkType,
-    MIMEType,
-    ClassLabel,
-    S3LinkAuth,
-    BoundingBox,
-    NumpyBinary,
-    BoundingBox3D,
-    GrayscaleImage,
-    LocalFSLinkAuth,
-    DefaultS3LinkAuth,
-    COCOObjectAnnotation,
-)
-
 from .model import Dataset
 
-__all__ = [
-    "Link",
-    "LinkAuth",
-    "DefaultS3LinkAuth",
-    "LocalFSLinkAuth",
-    "S3LinkAuth",
-    "MIMEType",
-    "LinkType",
-    "Binary",
-    "NumpyBinary",
-    "Text",
-    "Line",
-    "Point",
-    "Polygon",
-    "Audio",
-    "Video",
-    "Image",
-    "ClassLabel",
-    "BoundingBox",
-    "BoundingBox3D",
-    "GrayscaleImage",
-    "COCOObjectAnnotation",
-    "Dataset",
-]
+__all__ = ["Dataset"]
